handler/login.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/go-playground/validator/v10"
)

type Login struct {
	Service   LoginService
	Validator *validator.Validate
}

type ResponseBody struct {
	AccessToken string `json:"accessToken"`
}

func (l *Login) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	var b struct {
		Name     string `json:"name" validate:"required"`
		Password string `json:"password" validate:"required,min=8,custom-low,custom-upp,custom-num,custom-symbol"`
	}

	// jsonのデコード
	if err := json.NewDecoder(r.Body).Decode(&b); err != nil {
		RespondJSON(ctx, w, &ErrResponse{
			Message: err.Error(),
		}, http.StatusInternalServerError)

		return
	}

	err := l.Validator.Struct(b)
	if err != nil {
		RespondJSON(ctx, w, &ErrResponse{
			Message: err.Error(),
		}, http.StatusBadRequest)

		return
	}

	token, err := l.Service.Login(ctx, b.Name, b.Password)
	if err != nil {
		RespondJSON(ctx, w, &ErrResponse{
			Message: err.Error(),
		}, http.StatusBadRequest)

		return
	}

	rsp := ResponseBody{
		AccessToken: token,
	}

	RespondJSON(ctx, w, rsp, http.StatusOK)
}
